agent-runtime/pyproject.toml
[tool.poetry]
name = "agent-runtime"
package-mode = false
version = "0.1.0"
description = "LiveKit Agent Server"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
livekit-agents = "^1.4.1"
livekit-plugins-deepgram = "^1.4.1"
livekit-plugins-openai = "^1.4.1"
livekit-plugins-cartesia = "^1.4.1"
livekit-plugins-silero = "^1.4.1"
httpx = "^0.27.0"
pydantic-settings = "^2.2.0"
onnxruntime = "1.20.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.15.0,<0.16.0)"
]
[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["agent_runtime"]
